simulator/requirements.txt
# Simulator dependencies
google-cloud-pubsub==2.21.1
